chore(sidecar): ruff sạch toàn bộ context-engine (53 lỗi cũ → 0)
Auto-fix 44 (import order, unused noqa/vars, int-cast thừa, pairwise) +

sửa tay 11: contextlib.suppress thay try/except/pass, ClassVar cho

type_annotation_map, x thay ky-tu nhan, doi ten bien l/i mo ho.

Khong doi hanh vi; smoke import 8 module OK.

+++ apps/context-engine/src/directorai_context/modules/color.py
@@ -62,9 +62,9 @@ def _kmeans_dominants(image_bgr: np.ndarray, k: int = 5) -> list[DominantColor]:
         b, g, r = centers[i]
         dominants.append(
             DominantColor(
-                r=int(round(r)),
-                g=int(round(g)),
-                b=int(round(b)),
+                r=round(r),
+                g=round(g),
+                b=round(b),
                 fraction=float(counts[i] / total) if total > 0 else 0.0,
             )
         )
